+++ SwiftUI_Pull_to_Refresh/SceneDelegate.swift
@@ -21,6 +21,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Use a UIHostingController as window root view controller.
         if let windowScene = scene as? UIWindowScene {
             let window = UIWindow(windowScene: windowScene)
+			window.layer.speed = Float(UI.speed)
             window.rootViewController =
 //				UIHostingController(rootView: FitTextView())
 				AppHostingController(rootView: CitiesView())