+++ CHANGELOG.md
@@ -2,7 +2,21 @@
 
 * Doing
 
-	+ Blur effect performace (?)
+* Feature/Blur/0.1.1 - 0.1.3
+  
+	+ Slow animations
+		+ Added `UI.Speed`
+		+ Set `window.layer.speed` at start
+		+ Added `UIScrollView` speed extensions
+			+ `set(decelerationRate:)`
+			+ `set(pagingFriction:)`
+			+ Bouncing deceleration rate is unchanged
+
+* Feature/Blur/0.1.0
+
+	+ Scroll performance optimization
+		+ Added `BackgroundView` and `AlignedBackgroundView`
+			+ Extracted background (and blur) views to pre-baked assets
 
 * 1.8.0